🧹 Tidy: 陣痛履歴コンポーネントの分割とUtils/Libの統合
- ContractionHistoryをサブコンポーネントに分割し可読性を向上
- frontend/utilsをfrontend/libに統合しディレクトリ構成を整理
- 関連するインポートパスの更新とテストの修正
