+++ .jules/tidy.md
@@ -17,3 +17,12 @@
 ### アクション:
 - 新しいページを作成する際は、既存のレイアウトコンポーネントを積極的に利用する。
 - リソース操作（CRUD）を行うコンポーネントを見つけたら、既存のカスタムフック（`useRecordDelete` 等）が適用できないか検討する。
+
+## 2026-02-24 - [Fat Componentの分割とディレクトリ構造の明確化]
+### 学び: コンポーネントの責務分離とユーティリティの一元化
+- **Fat Componentの分割**: `ContractionHistory.tsx` にリストレンダリング、編集ダイアログ、削除ロジックが混在していた。`ContractionHistoryItem`（表示）と `ContractionEditDialog`（編集）に分離することで、親コンポーネントの見通しが良くなり、各コンポーネントの責務が明確になった。
+- **ディレクトリ構造の明確化**: `frontend/utils` と `frontend/lib` が混在し、開発者がどちらを使うべきか迷う状態だった。`utils` の内容を `lib` に統合することで、ファイルの配置基準が統一され、迷いがなくなった。
+
+### アクション:
+- 長大なリストレンダリングや複雑なダイアログを持つコンポーネントは、積極的にサブコンポーネントに切り出す。
+- 似た役割を持つディレクトリ（例: utils/lib, helpers/services）が乱立していないか定期的に見直し、統合する。
