+++ frontend/__tests__/growthUtils.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest"
-import { mergeData } from "@/utils/growthUtils"
+import { mergeData } from "@/lib/growthUtils"
 
 describe("growthUtils", () => {
     describe("mergeData", () => {
