+++ frontend/components/feeding/feeding-form.tsx
@@ -26,7 +26,7 @@ import { Input } from "@/components/ui/input"
 import { Textarea } from "@/components/ui/textarea"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Feeding, FeedingCreate, FeedingUpdate, FeedingType, BottleContentType, FeedingCompletion } from "@/types/feeding"
-import { buildFeedingPayload } from "@/utils/feeding-utils"
+import { buildFeedingPayload } from "@/lib/feedingUtils"
 import { FeedingTimerSection } from "./FeedingTimerSection"
 import { BreastFeedingFields } from "./BreastFeedingFields"
 import { BottleFeedingFields } from "./BottleFeedingFields"
