+++ frontend/components/growth/GrowthChart.tsx
@@ -14,7 +14,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Button } from "@/components/ui/button"
 import type { Growth } from "@/types/growth"
-import { generateWhoSeries, mergeData } from "@/utils/growthUtils"
+import { generateWhoSeries, mergeData } from "@/lib/growthUtils"
 import { useMemo, useState } from "react"
 import { format, subMonths, subDays } from "date-fns"
 import { Checkbox } from "@/components/ui/checkbox"
