+++ frontend/lib/growthUtils.ts
